+++ wemake_python_styleguide/visitors/ast/complexity/access.py
@@ -27,24 +27,12 @@ class AccessVisitor(BaseNodeVisitor):
         self._visited_accesses: Set[AnyAccess] = set()
 
     def visit_Subscript(self, node: ast.Subscript) -> None:
-        """
-        Checks subscript access number.
-
-        Raises:
-            TooDeepAccessViolation
-
-        """
+        """Checks subscript access number."""
         self._check_consecutive_access_number(node)
         self.generic_visit(node)
 
     def visit_Attribute(self, node: ast.Attribute) -> None:
-        """
-        Checks attribute access number.
-
-        Raises:
-            TooDeepAccessViolation
-
-        """
+        """Checks attribute access number."""
         self._check_consecutive_access_number(node)
         self.generic_visit(node)